Add travis test runner; update dependencies.
Note that the version of XeLaTeX is very old; we need to update the
hyperref package (and the ltxcmds package it depends on) and the
fontspec package (and the l3kernel and l3packages packages it depends
on).  Note that the new fontspec installation needs to be moved from
'latex' to 'xelatex' to properly override the default xelatex
fontspec.

The docs and sources have been removed from the tds installation
bundles in `texdeps` in order to avoid bloating our repository.

Change-Id: Ic8647430bd2f6255b1d3efb8b0313741047d2876