Use exiftool to strip resolution information from JPEGs to avoid a LaTeX error.
XeLaTeX will fail with a "dimension too large" error if the computed "actual
size" of the image is excessive.  For instance, if the JFIF X/Y resolution
value is set to "1dpi" and the image is 1200x800, \includegraphics will
fail.

Strip the X/Y resolution information with exiftool during image processing
to avoid this failure (we scale the image to fit the column widths anyway,
the EXIF dpi information is irrelevant).

Add a test case for this issue: the dewiki 'Papier' article contains an
image with overly-small resolution values.

Change-Id: I33e08508ce117b8843430bd23ac80ebaf321f1ca