Improve error-handling by avoiding use of the promise .done() method.
Creating a new deferred and then returning deferred.promise asynchronously
is somewhat of an antipattern: it breaks the call chain so thrown errors
don't get properly reported.  You can work around this by adding an
`otherwise` handler on the orphaned promise to call `deferred.reject(err)`
but it's better just to (eventually) return `deferred.promise` as part of
the normal promise chain.

Change-Id: I09b882c1b7130cc44661d2589fdf82eeef5d422d