Allow rendering from unzipped bundle
Allow the renderer to consume an unzipped bundle file.  This saves CPU
cycles when the bundle was created on the same machine.  Due to the
use of hardlinks, the bundle directory in this case must be on the
same file system as the temp dir -- use the system environment
variable TMPDIR to specify an appropriate directory if necessary.

Change-Id: Iad77b766a12654876a7c140220df9a37dc25f6c6