Use FAKESTYLES for FreeSerif.
The FreeSerif bold and italic fonts don't have as broad coverage as the
main FreeSerif font.  So use FAKESTYLES which synthesizes fake boldface
and italics fonts from the main font.  Quality of the font goes down
for headings, etc.  Sigh.  But it means that we don't get question
marks in Tamil, etc.

Bug: 28206
Change-Id: Ic543fe6c5544c79cd155404342f126e6641103b2