Ensure there is no race when cleaning up the gammalatex output directory.
When running on NFS -- or apparently on some tmpfs as well -- we can start
trying to clean up the output directory before the new file is actually
present in the directory.  That leads to EEXIST.

Fix submitted upstream in https://github.com/gammasoft/latex/pull/6 ;
in the mean time use a git hash in the `package.json`.

Change-Id: I828ee5794db2b9282ffd184833c81efbbbd26938