Use Lohit fonts when possible.
Use Lohit fonts for Bengali, Devanagari, Kannada, Malayalam, Tamil,
and Telugu. Remove old dependencies on fonts-nakula (for Devanagari)
and fonts-smc (for Malayalam).

Add a new 'nolatin' flag for a script font, which indicates that (as
for the Lohit fonts) you need to manually switch back to the latin
font for characters in the Latin-1, Latin Extended A, and Latin
Extended B codepages (otherwise you get tofu).

Bug: 28206
Bug: 68922
Change-Id: I19f2ba69b19e4533fedc139587f182e582231c46