Don't crash on `;#`.
The article [[en:1988 in film]] contains `;#` -- which is a typo, what
is meant is `; #` (that is, the definition list term is "#").  But `;#`
generates `<dl><dt><ol><li>...` which confuses LaTeX.  Generate a
`\parbox` in this case so that even this broken markup works.

(Our test for whether the definition term is "too complicated" is
hacky.  A better one would be welcome.)

Change-Id: Ic45aff072660f4ad46b622fe90e4796c57eb16ab