T93008: Use {{DISPLAYTITLE}} or subpage title where appropriate.
We now work harder to find the best title for an article in the
generated PDF.  We first use `item.displaytitle` if set; this would be
a user's explicit override.  If that's not set, we use the value of
the `{{DISPLAYTITLE}}` magic word, if present.  If that's not present,
we use the subpage title if subpages are enabled in the namespace and
this is not a single-item collection.  Otherwise we fall back to the
`item.title` as before.

NOTE that this does not affect the page title for single-item
collections, since that comes from metabook.title, which is set
by the Collection extension.

Change-Id: I216eff064ee251d17020a5a4586a70149eaea431